requirements.txt
PyOpenGL
glfw
Pillow==9.5.0